#381 change pandas tests order
